# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OCR Enterprise System API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalJobs: number;
  completedJobs: number;
  pendingJobs: number;
  processingJobs: number;
  failedJobs: number;
  avgProcessingTimeMs: number;
  avgConfidenceScore: number;
  highQualityCount: number;
  mediumQualityCount: number;
  lowQualityCount: number;
  totalUsers: number;
}